Use printf() for everything
This will show you the bug - you try to print a C-style string starting
at nullptr. The std::cout seems to not like that; printf() handles this
better.

Anyway, it would be a good idea not to print null characters (ie, the \0
byte) to standard output in cooked mode. Some terminals don't like it.